requirements.txt
aiohttp==3.7.4.post0
alabaster==0.7.12
amqp==5.0.6
anyio==2.0.2
anyjson==0.3.3
appnope==0.1.2
argcomplete==1.10.0
argon2-cffi==20.1.0
asgiref==3.2.10
async-generator==1.10
async-timeout==3.0.1
attrs==20.3.0
Babel==2.8.1
backcall==0.2.0
beautifulsoup4==4.9.3
billiard==3.6.3.0
bleach==3.2.1
boto3==1.16.4
botocore==1.19.4
bs4==0.0.1
celery==5.0.5
certifi==2020.6.20
cffi==1.14.4
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
commonmark==0.9.1
ConfigArgParse==1.2.3
cryptography==3.4.7
curlify==2.2.1
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
diff-match-patch==20181111
Django==3.1.2
django-admin-rangefilter==0.6.0
django-appconf==1.0.4
django-better-admin-arrayfield==1.1.0
django-bootstrap-pagination==1.7.1
django-cacheops==5.0
django-ckeditor==5.9.0
django-daterangefilter==1.0.0
django-debug-toolbar==2.2
django-filter==2.4.0
django-imagekit==4.0.2
django-import-export==2.1.0
django-inline-actions==2.3.0
django-js-asset==1.2.2
django-json-widget==1.0.1
django-lazysignup==2.0.0
django-mathfilters==1.0.0
django-modeladmin-reorder==0.3.1
django-nested-inline==0.4.2
django-object-actions==3.0.1
django-redis==4.12.1
django-report-builder==6.3.2
django-request-logging==0.7.2
django-searchable-encrypted-fields==0.1.9
django-sendgrid-v5==0.8.1
django-sequences==2.5
django-ses==1.0.1
django-simple-history==2.12.0
django-solo==1.1.1
django-storages==1.10.1
django-summernote==0.8.11.6
djangorestframework==3.12.1
djangorestframework-api-key==2.0.0
dnspython==2.1.0
docutils==0.16
docx2txt==0.8
drf-extra-fields==3.1.1
EbookLib==0.17.1
entrypoints==0.3
et-xmlfile==1.0.1
ExifRead==2.3.2
extract-msg==0.23.1
Flask==1.1.2
Flask-BasicAuth==0.2.0
funcy==1.14
future==0.18.2
gevent==21.1.2
geventhttpclient==1.4.2
greenlet==1.0.0
gunicorn==20.0.4
haversine==2.3.0
idna==2.10
image==1.5.33
imageio==2.9.0
imagesize==1.2.0
IMAPClient==2.1.0
importlib-metadata==1.6.0
ipykernel==5.4.3
ipython==7.28.0
ipython-genutils==0.2.0
ipywidgets==7.6.3
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.18.0
Jinja2==2.11.2
jmespath==0.10.0
joblib==0.17.0
Js2Py==0.71
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.11
jupyter-console==6.2.0
jupyter-core==4.7.0
jupyter-server==1.2.1
jupyterlab==3.0.4
jupyterlab-pygments==0.1.2
jupyterlab-server==2.1.2
jupyterlab-widgets==1.0.0
kiwisolver==1.3.1
kombu==5.0.2
linkhub==1.2.0
livereload==2.6.3
location==0.0.7
locust==1.0.2
logging==0.4.9.6
lunr==0.5.8
lxml==4.6.1
Markdown==3.3.2
MarkupPy==1.14
MarkupSafe==1.1.1
matplotlib==3.3.3
matplotlib-inline==0.1.3
mistune==0.8.4
mkdocs==1.1.2
mkdocs-encryptcontent-plugin==0.0.13
mkdocs-material==6.1.6
mkdocs-material-extensions==1.0.1
msgpack==1.0.0
multidict==5.1.0
nbclassic==0.2.6
nbclient==0.5.1
nbconvert==6.0.7
nbformat==5.0.8
nest-asyncio==1.4.3
networkx==2.5
nltk==3.5
notebook==6.2.0
numpy==1.19.4
odfpy==1.4.1
olefile==0.46
opencv-python==4.4.0.46
openpyxl==3.0.3
packaging==20.4
pandocfilters==1.4.3
parso==0.8.1
pdfminer.six==20181108
pexpect==4.8.0
pickleshare==0.7.5
piexif==1.1.3
pilkit==2.0
Pillow==8.0.1
pipdeptree==2.0.0
popbill==1.20.0
prometheus-client==0.9.0
prompt-toolkit==3.0.10
psutil==5.7.0
psycopg2-binary==2.8.6
ptyprocess==0.7.0
pycparser==2.20
pycrypto==2.6.1
pycryptodome==3.12.0
pycryptodomex==3.10.1
Pygments==2.7.2
pyjsparser==2.7.1
pymdown-extensions==8.0.1
pyOpenSSL==20.0.1
pyparsing==2.4.7
pypng==0.0.20
PyQRCode==1.2.1
pyrsistent==0.17.3
python-barcode==0.11.0
python-dateutil==2.8.1
python-http-client==3.2.7
python-pptx==0.6.18
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzbar==0.1.8
pyzmq==20.0.0
qtconsole==5.0.1
QtPy==1.9.0
redis==3.5.3
regex==2020.11.11
requests==2.24.0
s3transfer==0.3.3
scikit-image==0.18.1
scipy==1.5.4
Send2Trash==1.5.0
sendgrid==6.4.3
sentry-sdk==0.19.3
six==1.15.0
slackclient==2.9.3
slacker==0.14.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcontainers==2.2.2
soupsieve==2.0.1
SpeechRecognition==3.8.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.4.1
starkbank-ecdsa==1.0.0
tablib==2.0.0
terminado==0.9.2
testpath==0.4.4
tifffile==2021.1.14
tornado==6.1
tqdm==4.51.0
traitlets==5.0.5
typing-extensions==3.7.4.3
tzlocal==1.5.1
urllib3==1.25.11
vine==5.0.0
watchtower==1.0.6
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlwt==1.3.0
xmltodict==0.12.0
yarl==1.6.3
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
